#994 - sensor slope and intercept are now stored as floats.

+++ grails-app/domain/au/org/emii/aatams/Sensor.groovy
@@ -19,12 +19,12 @@ class Sensor extends Tag
     /**
      * Calibration slope.
      */
-    Integer slope
+    Float slope
     
     /**
      * Calibration intercept.
      */
-    Integer intercept
+    Float intercept
 
     static constraints =
     {
